♻️ Refactor: any 타입을 이용한 임시 마이그레이션 완료
