--- src/folders_mockData.json
@@ -1,84 +1,0 @@
-{
-	"data": [
-		{
-			"id": 14,
-			"created_at": "2023-06-04T20:59:39.293024+00:00",
-			"name": "⭐️ 즐겨찾기",
-			"user_id": 1,
-			"favorite": true,
-			"link": {
-				"count": 0
-			}
-		},
-		{
-			"id": 24,
-			"created_at": "2023-06-18T13:33:09.993192+00:00",
-			"name": "신규 폴더1",
-			"user_id": 1,
-			"favorite": false,
-			"link": {
-				"count": 0
-			}
-		},
-		{
-			"id": 307,
-			"created_at": "2024-01-12T15:08:12.873779+00:00",
-			"name": "새로운폴더스",
-			"user_id": 1,
-			"favorite": false,
-			"link": {
-				"count": 0
-			}
-		},
-		{
-			"id": 433,
-			"created_at": "2024-01-14T12:01:27.040783+00:00",
-			"name": "유담이와 아이들",
-			"user_id": 1,
-			"favorite": false,
-			"link": {
-				"count": 0
-			}
-		},
-		{
-			"id": 441,
-			"created_at": "2024-01-14T12:40:56.383949+00:00",
-			"name": "안 없어져",
-			"user_id": 1,
-			"favorite": false,
-			"link": {
-				"count": 0
-			}
-		},
-		{
-			"id": 457,
-			"created_at": "2024-01-14T13:47:32.813674+00:00",
-			"name": "또",
-			"user_id": 1,
-			"favorite": false,
-			"link": {
-				"count": 0
-			}
-		},
-		{
-			"id": 480,
-			"created_at": "2024-01-15T11:03:07.802566+00:00",
-			"name": "추가",
-			"user_id": 1,
-			"favorite": false,
-			"link": {
-				"count": 0
-			}
-		},
-		{
-			"id": 513,
-			"created_at": "2024-01-18T15:55:27.603912+00:00",
-			"name": "string2",
-			"user_id": 1,
-			"favorite": false,
-			"link": {
-				"count": 0
-			}
-		}
-	]
-}
